Add FixedSizeStack helper class
Used in KdTree and BVH intersection methods instead of plain array and
accompanying counter.

It looks like GCC 4.9 is not capable of optimizing this to the same thing
as before. Stack size is not kept in a register all the time. This means
that intersection is a bit slower (boing ball scene rendered in 12000x8000
with the CLI application goes from ~13.3s to 13.4s). Nevertheless I think
this change is worth the extra readability it brings. And there really is
no reason why GCC should not optimize it to the same thing.

Another thing to note is that elements are not cleared in pop(). Doing
array_[size_] = Type() clears the element with zeros. Even if a constructor
that does not set any value is defined. I think GCC is in violation of the
C++ specification here. An element can still be cleared by calling
stack.top() = Type() before doing stack.pop(). All things considered,
perhaps it is not a bad thing to require the caller to be explicit.