Stage VersionInfo from main repo
Use Meson's vcs_tag() function to fetch a version string from git.

The whole point of this is to only have the version in git and have it be
automatically updated to tag[-<num commits after tag>-g<hash>][-dirty] when
making changes. See 'git help describe'.

The idea with setting fallback for vcs_tag() to '<git describe fallback>'
is to have the release script replace it when making a release archive.

Handling building from an archive was perhaps a bit easier in the Make
release target of the main repo, where a file was just shipped with the
version string in it. Will fix so this works with Meson as well but no time
for it now. Want to get this pushed. (The point of today's exercise was
to experiment with vcs_tag() and find out what would be easy/hard to port
and that has been accomplished. :)