Add forgotten Variant to Quaternion test
Variant to Quaternion is actually used since Transform from Quaternion is
allowed in scene format. Not a user friendly way to specify rotations but
it is mainly there to allow for e.g. writing a Blender exporter in Python,
where I would assume it is easier (and better!) to have the rotation
transferred to Rayni as a quaternion instead of as Euler angles, for
instance.