Fix unwanted conversions in DurationFormatter
By using underlying representation of std::chrono durations and by using
an std::uint8_t for setting precision of seconds.

Stream precision is a bit weird since it takes an int but it is not exactly
clear, I think, in the standard what a negative precision means. Both
libstdc++ and libc++ seems to interpret it as default for string streams.
But instead of relying on that and having to have to handle it when
determining width etc, use an std::uint8_t which is OK for the implicit
conversions in DurationFormatter::format().