Add Meson option for building with -Ofast and -march=native
Could be done with CXXFLAGS= and LDFLAGS= when invoking meson but that
is error prone. Will also port other build options from Makefile based
build system in main repository soon. Better to have everything in the
build system instead of describing it in a document and letting the user
copy paste.