Add option for building with libc++ instead of libstdc++
The Make based build system in the main repository does this by default if
Clang is used to build. This is currently not possible with Meson though
since the GTest dependency uses prebuilt .so files if available (which they
are on Arch). Uninstalling the gtest package and checking out GTest in a
location where Meson looks for its source also does not work since the same
set of warnings are used as for the rest of the project. Not sure what the
best solution is. Check this in for now and come back to it later.