Add Meson option for building with memory sanitizer
Note that all libraries must be built with memory sanitizer for it to give
any useful reports. See the memory sanitizer documentation. Makes it a bit
unpractical to use. Nice when it works though.

Immediate segfault when building with libc++ as it is now. Probably
because of -lsupc++ required for static thread_local variables. Needs to
be implemented properly in libc++ so dependency on GCC's libsupc++ can
be removed.

Pass -fsanitize-memory-track-origins to track where memory is allocated.
1.5-2.5 slowdown but more informative reports.

Pass -fno-omit-frame-pointer for stack traces.