Include related header first in .cpp files
Have refrained from this in the past since it is an extra rule to follow
that is not obvious. Changed my mind now that clang-format sorts includes
and has logic to handle this. It is not perfect (only sorts in blocks) but
hopefully it will be improved in the future.