Configure clang-tidy readability-identifier-naming check added in 3.8
Each identifier type Foo has a FooCase, FooPrefix and FooSuffix
configuration variable. FooCase is set to "any case" by default and
FooPrefix and FooSuffix are set to the empty string. No identifier type
should have a prefix or suffix in the current style, so none are set in
.clang-tidy.

Configure each identifier type explicitly. There is logic for not having
to configure each type (e.g. if struct is not set, fall back to class etc.).
This is not documented though and the logic for it is not always clear.
There is also a problem with not being able to distinguish between
"any case" and "not set". Reevaluate if these deficiencies are ever fixed.