Allow for enum_from_value() to take a std::initializer_list
Will be used to remove some static const variables in TGAReader.

Both GCC and Clang generate the same code when exchanging a std::vector
with a std::initializer_list in the case of TGAReader.

std::initializer_list does not have cbegin() and cend() so change to
begin() and end() (no difference since enum_values is const).