Use range-based for in Matrix4x4::swap_columns()
Clang 3.8 generates the same code.

GCC 6.1.1 however generates worse code (instead of unrolling and doing
simple move operations it generates a loop with jump instructions). Bad
of GCC but hope it is fixed soon in a later version. Column swapping is
currently not done in any performance critical rendering code.