Make it possible to convert a Variant to an Image
Only support map with "path" as key for now. Could also allow for base85
encoded binary data etc (e.g. to allow for generating image textures in
Python) but do this at a later point. This is all that is needed to move
ImageTexture creation out of SceneReader for all current test scenes that
use image textures.