libjpeg-turbo 1.5.0 ships a pkg-config file, require that version
Previously distributions patched this in by themselves. Some of them, e.g.
Debian and Ubuntu, in a buggy way. Requiring 1.5.0 will still not work on
Debian and Ubuntu since the version string is empty. But this will at least
make it clear that a version with an upstream pkg-config file is required
and it will start working when they remove their broken patches.

Relying on a libjpeg-turbo version number instead of a libjpeg one should
not be any problem since libjpeg-turbo specific API is used and it is the
version all major distributions use anyway.