RFC 7159 no longer require JSON documents to be arrays or objects
Lifts restriction that was in RFC 4627 (which was the latest available RFC
for JSON when the parser was originally implemented in Rayni).

Move tests for leading and trailing characters down since they previously
used an array/object root since all other tests did that as well. Now it is
better to test after everything else so we know that the test values can be
parsed. Also add a couple of more trailing chars tests.