Add BigEndianReader for reading big endian values from data
Allow for reading integer values of various sizes and fixed point values
with arbitrary denominator that are converted to real_t:s.

Also allow for extracting a group of values that will be stored in a
std::array. This in turn can be used to to convert groups of every x:th
element to a type that has a constructor that takes a std::array of size x
as an argument. See the BigEndianReader.Read*VectorArrayConstructor tests
for examples. This will be utilized to extract Vector3, UV and triangle
indices in a simple and elegant manner when reading meshes generated in
Python.