Return reference instead of pointer in all T::get_from_variant
Should not be possible to return null.

Allows for changing *T::get_from_variant(v) to v.to<const T &>().

Keep Variant::to() for T * but implement it with Variant::to() for T &.
This simplifies Material that stores a Texture pointer since it should
be possible to have it set to null if no Texture is used in the Material.

Comment about ownership is removed since it should be obvious now that a
reference is used instead of the pointer in the implementations.