Add back libcxx build option
First added in 98138e1e for Meson. Removed in 9ebc5fde to mimic behavior of
Make based build system in main repository.

Add it back now since libc++'s llibc++experimental is not readily
available in most Linux distributions and it does not look like it will be
in the near future (if it ever happens). C++17 will probably be readily
available before that. Do not want to fiddle with build files every time
test builds with Clang are done until then though.