NOLINT const_cast for execvp()
Not much to do. POSIX requires non const due to limitations of ISO C.
See RATIONALE section in:

http://pubs.opengroup.org/onlinepubs/009604499/functions/exec.html

"The statement about argv[] and envp[] being constants is included to
make explicit to future writers of language bindings that these objects
are completely constant. Due to a limitation of the ISO C standard, it
is not possible to state that idea in standard C. Specifying two levels
of const- qualification for the argv[] and envp[] parameters for the
exec functions may seem to be the natural choice, given that these
functions do not modify either the array of pointers or the characters
to which the function points, but this would disallow existing correct
code. Instead, only the array of pointers is noted as constant."