Move Pipe out of src/lib/system/command.cpp
To a linux/ subfolder for now since pipe2() (Linux specific) is used.
Have other Linux specific stuff and do not want to compromise with
subpar API:s when better ones are available. Will move stuff
around/refactor when/if another POSIX OS ever becomes interesting.

Keep write file descriptor duplication in command.cpp. Do not want to
expose it as a method in Pipe now that it is accessible to the outside
world.