modernize-use-auto has started to warn about casts in clang-tidy 4.0
Already used auto in a couple of other places when casting.

The check now also warns about calling templated functions that behave as
a cast. Not sure about this one though. How does it know for sure that it
is how e.g. Variant::get() is supposed to behave? Only have one occurrence
of this in src/lib/color.cpp so remove temporary variables instead of using
auto.