Add helper methods to Variant for getting key/index in parent
A bit ugly with all the asserts() but...

Instigated by performance-inefficient-string-concatenation added in
clang-tidy 4.0 warning about using std::string's + operator in for loops.
False positive in this case since it is only used once in the return
statement in the loop. But too lazy to file a bug for this. It may also be
hard to catch all cases similar to this in clang-tidy so leave it be.

Could be argued that the helper methods make sense even if clang-tidy did
not complain in this case.