Move MainLoop::TimerData out of header to .cpp file
Will add support for monitoring file descriptors soon and header is
becoming quite large. Really no need to have TimerData and similar class
for FD:s in header when they are stored in a std::shared_ptr.

Id and Id constant used for empty timers are moved to MainLoop so they
can be used in header for Timer.

Move comments about std::map vs std::unordered_map and why a recursive
mutex is required to start of class definition and also explain why a
std::shared_ptr is used.