Add new clang-format 5.0 options and run on all source files
Want to change AlignEscapedNewlines to None now that it is no longer a bool
but it crashes in 5.0. Test at some later point with a dot release and file
a bug if it is not fixed.

BreakConstructorInitializersBeforeComma was replaced with
BreakConstructorInitializers . Will modify it in following commit now that
it is an enum instead of a bool with a value that is what my original style
used.

FixNamespaceComments does not seem to work. Buggy?