clang-format 5.0 learned to format initializer lists the way I want, use it
04d0d7d664032efc5ddd00ee2c0f8c2329f861af in main repo modified how
initializer lists were formatted since there was now way to make
clang-format line break after colon and put each variable on a separate
line.

BreakConstructorInitializersBeforeComma (a bool) was replaced with
BreakConstructorInitializers (an enum) in 5.0 and it now possible to do
this so revert back to my original style.

Now it is easier to read initializer lists again when each variable is
always on a separate line. \o/ (More noticeable in main repo with a lot
more code. :)