Add a trailing _ to all instance variables.
Because readability-identifier-naming check in clang-tidy 5.0 no longer
allows an optional _.

Only had _ for instance variables where they would conflict with getters
(getter for variable foo_ is foo()). This is no longer possible if a _
is not appended to all instance variables.

Not really fond of encoding defails about a variable in its name. But
readability-identifier-naming is a check that is too valuable to have
disabled. So follow its whims and peculiarities. (Not sure that not
allowing a _ anywgere is not a bug. Have looked at the code in
clang-tidy and understand case check was added for but too lazy to do
anything about it. Not the end of the world and appending a _ is popular
due to it being in the Google style guide.)