Set readability-function-size.ParameterThreshold added in clang-tidy 5.0
Only a couple of methods/functions with 8 arguments. One in KdTree::Builder
used when splitting events and objects along an axis. Could reduce it to 4
by adding a struct for left/right side. But there are also a couple of
GTest predicate functions that take 8 arguments (value and string
representation for each parameter builds up fast) so just set it to 8.
Better than nothing and a power of 2 number. ;)