cppcoreguidelines-special-member-functions can still not be enabled
Due to violations in Google Test. Set AllowSoleDefaultDtor added in
clang-tidy 5.0 anyway and change reason for why it is disabled.

A bit frustrating that Google is not more interested in fixing lint
problems in Google Test since they seem to be large proponents of
clang-tidy. One would assume they use it a lot in their own projects and
have the same problems. Oh well, maybe it is time to have another look
around for a more modern, up to date and better maintained test framework
for C++ again. (This is far from the only issue I have with GTest. *sigh*)