Task vector passed to ThreadPool::add_tasks() should be empty afterwards
Do not rely on this anywhere at the moment. But it is a reasonable
expectation, I think, since it is moved.

Clear it while not holding the ThreadPool mutex.

Could unlock it before notifying as well. But not really related to this
change. Also most likely a premature optimization since many (most?)
implementations avoid the spurious wake up of receiving threads that could
be done because condition is notified while holding the lock.