Update for clang-format 6.0
Remove AfterObjCDeclaration from BraceWrapping. Not interested in ObjC.
(Accidentally added when it was new in 3.8.) AfterExternBlock is new in
6.0 . Do not have any extern blocks but might as well continue being
explicit. (Have been bitten by default values changing before...)

Set FixNamespaceComments to false. Added in 5.0 but did not work. Works in
6.0 but thinking a bit more about it this is just clutter similar to
commenting end braces in large if/else blocks. If it is hard to keep track
of what start brace an end brace belongs to it is a sign that things should
be broken up. For namespaces, that probably means creating smaller files.

6.0 fixes a bug with lambda captures and init expressions. Extra space is
no longer added. See https://bugs.llvm.org/show_bug.cgi?id=19986 .

Include block grouping finally added! Set IncludeBlocks to "Regroup" so
that header is placed first. Have tried to follow this but missed it in a
few places (tooling ftw!). Never tried to follow it in tests but now that
it can be enforced with clang-format, do it there as well. (Catches missing
includes in headers that do not have a corresponding .cpp file.)

Indentation of include directives have also finally been added. Enable it.
Think it looks a bit weird with the level of indentation (have used one
space historically) but I guess it makes sense to follow the rest of the
code and indent with tab.