Modify clang-tidy NOLINT comments to only disable specific checks
Feature added in clang-tidy 6.0. It also learned NOLINTNEXTLINE. Use it
in a couple of places where line otherwise would be very long.

Also add comments explaining why in a couple of places that previously
missed it. (Should perhaps have been a separate commit...)

misc-use-after-move was also renamed to bugprone-use-after-move in 6.0. Not
noticed before when it was only a comment.