Build in release by default
Release is mostly what is wanted. What the old build system did. Want it to
be easy to do what is done 99% of the times a build is created.

Very seldom debugging is needed. When a scene is rendered incorrectly,
trying to render it in a debug build is often too slow and some other
means of debugging has to be used. It is also very often the case that the
error does not reproduce in a debug build due to different floating point
operations being used.