Build with C++17
Have a list of improvements that can be done, but only enable for now. Will
make changes piecemeal going forward.

libc++ support is probably broken or if not, it will break when
std::experimental::filesystem is changed to std::filesystem in a following
commit. Do not want to wait for libc++ to catch up. Onwards!