Fix a couple of clang-tidy modernize-use-auto warnings
Not sure why they started to show up now all of a sudden. It is not due to
moving to C++17 (shows up in earlier commits now as well) and clang-tidy
has not been updated since 6.0.0 was released (at which time I checked all
source for sure).

Do not like clang-tidy warning in these places (e.g. not obvious that
return value of Variant::get<T>() is T). Not worth adding NOLINT:s and
there are other similar places where it does not warn. Oh well...