Temporarily change locale in string to float tests
To test that functions are not affected by the global locale. Which is the
main point of having these functions, so good if it is tested...

std::locale::global() changes the global C++ locale as well as the global
C locale.

Perhaps not Windows friendly but that is a problem for later when Windows
support is added again. Should hopefully just have to use another locale name.

May also be that the test locale is not installed. Do not think there is a way
to detect this though. I think tests will succeed regardless, they just will not
test that the global locale is not used.