Replace string_to_long() with std::from_chars() (added in C++17)
The reason for std::from_chars() being added to C++17 is the exact reason
for why string_to_float/double() exist. (Meant to be used for file format
parsing and the like, where the current locale should not matter.)

Only use it for integers though since libstdc++ has not implemented parsing
of float and doubles in GCC 8.1.0. libc++ 6.0.0 has not implemented it at
all but will hopefully have some support soon (do not build with libc++
that often since it was removed from Arch). Add a TODO for floats and
doubles. Also have to hold off with using std::string_view. See TODO.

std::from_chars() behaves different than std::strtol() when it comes to
space and '+' at the beginning of the string. This is OK though since it is
up to the file format readers to strip space etc when parsing (which they
currently do in all cases where string_to_number() is called). Old behavior
was mostly tested for since it was how std::strtol() behaved.

It looks like there is a slight performance increase with this change.
Parsing the Stanford Buddah, Bunny and Dragon in the PLY file format
with the "format" property set to ASCII goes from ~3.3s to ~3.1s. There
are roughly 8 * 10^6 integer strings (used for triangle indices) parsed in
this case.