Generate Base85 decoding table at compile time
std::array's index operator was modified to be constexpr in C++17, so now
base85_generate_decoding_table() can be constexpr as well.

Can not use std::fill() though (since it is not constexpr, fixed in C++20)
so use {} to initialize to zero instead. A bit more unclear maybe, but
shorter... if someone wonders why it is there and removes it, both GCC and
Clang will fail to compile (uninitialized variable in constexpr).