Use std::string_view for string_to_number()
Will modify WavefrontObjReader to use function for splitting face string to
std::string_view:s and to use string_to_number() instead of std::strtoul()
in following commits. Requiring a std::string results in unnecessary data
copying that is noticeable when reading large object files. ~4.5s instead
of ~4s when reading 2 * 10^6 faces.

As for why removing std::string creation when reading faces help compared
to reading vertices and normals, it is probably due to more copying.
1 * 10^6 vertices and normals each and 3 floating point strings for each is
6 * 10^6 floating point strings. 2 * 10^6 faces with 6 indices for each
(triangles and only vertex and normal indices, no uv) is 1.2 * 10^7.

Tried with replacing argument with a template type so both std::string and
std::string_view would be handled automatically. That does not work with
string literals though. Could add special handling for this but since the
long term plan is to use std::string_view and slowdowns where not that
noticable, just do it now.