Replace PLYReader int to float/double conversions with methods in BinaryReader
Only used when reading PLY object files and original idea was to not have
these in BinaryReader since the conversion is ugly and do not want to
implement "real" IEEE 754 float binary interpretation. But, considering
that every interesting platforms under the sun uses IEEE 754 and
ply_reader.cpp is one of the largest files in the project at ~550 lines,
add this functionality to BinaryReader. It might be useful for some other
file format in the future.

Also be clear with that it is IEEE 754 by naming the methods accordingly.
The static asserts is probably not a guarantee for checking that the union
hack actually works compile time, but... c'est la vie!