Remove comment about checking .gz magic in download scene data script
Will not check magic for .rar files (will add support in following commit).
Not worth the hassle.

We have checksumed the files so should not be a problem in practice
security wise. Was just being a bit too forward thinking and having in
mind that _extract() will probably be copy pasted to other projects in the
future. Will probably never do anything about it if a gzip.is_gzipfile() is
not added anyway.