Remove objects/ dir in scene download dir
Enough to put Stanford objects directly in stanford/ instead of in
objects/stanford/ . Larger scenes will be put in <scene>/, do not want to
sort out objects and textures into separate folders. If there are separate
objects just put them in <name of source>/ or the download root.

Also remove variable for Stanford download path in scene data download
script as well. Not worth to have it. If a typo is made, the file(s) will
not load since they do not match paths in the JSON or Python file. Too lazy
to split this out into a separate commit.