Add support for RAR archives to download scene data script
By using the unrar command.

Python does not have any RAR support builtin (like for Tar and Zip).
python-rarfile exists but is not available in Arch. Packaged for Fedora
and Debian though so considered using it. Turns out that it is implemented
by just invoking unrar though so might as well do it directly since it is
not that complicated.