Add Variant::to_bool() so Variant::get<bool> works
Variant::get<T>(const std::string &key) does Variant::get(key).to<T>().

Probably did not add it in the first place since it is better to use an
actual bool Variant than to use a Variant with a numeric value (and have
false be 0) and to_* means "convert to". Have as_* for getting value
without doing any conversion. Could maybe special case bool in
Variant::get<T>() but do not think that is nice. May also end up adding
conversions from numeric values to bool in the future.

Have not had a need for a boolean value in scene file format until now.
Will be used in following commit to add a boolean "calculate_normals" key
to file objects.