Enable cppcoreguidelines-pro-type-union-access
Now that C++17 is available and union is no longer used in Variant.

A union is used in BinaryReader for type punning but just NOLINT it.

The bug mention in the removed comment (unions from system headers) is
still valid and triggers when accessing sigaction.sa_handler in CLI
application. NOLINT and move LLVM bug link to comment there.