Make ListnerList::ListenerBase assignment and move ctors/ops public
clang-tidy's modernize-use-equals-delete warns about deleted member
functions not being public. It is not documented why and code does not
mention it either. Presumable it is to give compiler a chance to output
better error messages if the deleted functiones are used by mistake... ?

But it also makes sense, I think. Used indirectly through subclass in
"public context". Constructor and destructor should still be protected
though since it is not meant to be instantiated directly.

Move friend declaration down to private section as well instead of keeping
it at top.