Stage TriangleMeshData
Only have a couple of simple tests for normal calculation.

Unsure if TriangleMeshData::from_variant() will be kept long term. Original
idea was to be able to generate meshes in Python. Only have a single scene
that does that because it is so convoluted to use numpy. Easier and faster
to generate meshes in C++. But, want to keep Variant conversion around
until a decision has been made on how to convert Blender scenes. (Will it
be done in Python or C++? Maybe a Python tool that generates a Rayni scene
and writes PLY files for meshes? Leaning towards last option at this
point... but implementation is far away.)