Simplify readability-identifier-naming config in .clang-tidy
Decided for explicitly setting every value when configuring, see
83e48aeaf27af3c14383 . Even though deficiencies mentioned in that commit
still stand (not documented etc.) it is possible to reduce size of
.clang-tidy so much, so have gone throug the check's source code and
deduced what can be removed.

Have also tested so that wrong naming is still detected correctly for all
types where config has been removed. Must do since behavior is that if
there is no matching config for type, do not check it. *sigh*

Wish it was documented and/or was a way to ask clang-tidy to collapse
configuration when it is similar for different type.