Tell misc-non-private-member-variables-in-classes to ignore POD structs
With helper constructor(s).

This will potentially allow classes (non POD structs) to have public
member variables as well. misc-non-private-member-variables-in-classes
should probably ignore constructors in PODs, but it does not. Too lazy to
file a bug. Perhaps allowing classes to have public variables if they are
all public is a good thing even.