NOLINT readability-convert-member-functions-to-static *Reader::read/write_file()
With a TODO about how it will be turned into a function. Reason for having
class now is mostly for following the pattern:

try
{
        image = FooReader().read_file();
}
catch (const FooReader::Exception &e)
{
        ...
}

Long term plan is to remove *Reader classes (except for in src/lib/io) and
just have free functions when Result (that contains value or descriptive
error message, accessing value if there was an error will throw) is
finished. Want to change this throughout the src/lib/file_formats/ folder
though, which is a large change, and not 100% happy with Result at the
moment so have it on a branch. Just NOLINT and TODO for now to get
clang-tidy to not complain.