Add FixedSizeStack::pop_or() to simplify BVH intersection code
Modify pop() to move out value. Have a vague memory of looking at assembly
for KdTree intersection code when testing long ago. GCC generated worse
code then in KdTree::intersect() so did not do it. Now it looks like GCC
ignores the unused return value and assembly looks the same, so do it now
instead of only moving in pop_or(). It makes sense to move in both.

For BVH::intersect() it does not matter since the type stored in the stack
is trivial (a node pointer, while KdTree has a struct).