Take error bound into account when intersecting AABB
See reference for details (online free version of book available since a
year or so back).

C++ epsilon vs MACHINE_EPSILON is a bit confusing. In most places outside
of academia, and thus also in C++, a definition that is twice as large
is used. In PBRT the academic version is used in this case. (Think it makes
sense. Was a while since I read the Managing Rounding Error section in
detail but... want error bound to be as small as possible...)