Pass std::chrono::nanoseconds to non template duration_format()
Simplifies expressions in if statements of duration_format(). The
expressions are also easier to read now, I think. Less code is inlined as
well.

For some reason moving the float ss value from the header leads to less
precision so that one of the DurationFormat.SecondsPrecision tests fail:

../src/unit_tests/lib/string/duration_format.cpp:53: Failure
Expected equality of these values:
  "34.51"
  duration_format(34s + 505ms, {.seconds_precision = 2})
    Which is: "34.50"

Changing ss to a double instead of a float fixes this.