Use Result in MemoryMappedFile
Will start using MemoryMappedFile in BinaryReader and TextReader in
following commits and this removes need for try/catch when opening a file.
(Will replace need for try/catch in other places by using Result instead
of exceptions in the future for more clarity/ease of understanding.)

Remove MemoryMappedFile constructor that takes a file name since there is
no way to report an error directly from the constructor wihtout exceptions.
Could change it to log an error message and require user to check if
mapping succeeded or not afterwards (by e.g. checking if data() is null).
But more error prone, so just remove it.