Simplify lib/version_info.h
By removing a separate header for "git describe" .in file and just using
constexpr constants instead of static methods.

For guaranteeing that generated headers are generated before they are used
outside of a dependency they must be in sources of declare_dependency(),
so do that for lib_dep in src/lib/meson.build (intermittent build failuers
in src/ui otherwise). Include config_header as well even though it is only
used in src/lib at the moment.

Remove unit tests. Was a bit pointless before and even more so now.