Remove std::error_code from Error
And just use it directly in Error() to create error message. If it will
ever be needed to get more information about an error it is easy to add it
back.

Want to reduce size of Error. It is still large (32 bytes in 64-bit). May
store it or message string as pointer in future. Still undecided.