Use Result for file_write()
Tests in src/unit_tests/ that use file_write() to write temporary files to
test file reading is modified to wrap all file_write() calls in an
ASSERT_TRUE(). Was not needed before since exception was thrown.
ASSERT_TRUE() will have the same effect. Test will be aborted. It is not
really the same thing since "test failed" will be reported instead of
"fatal error", but it does not really matter. Easy to see what went wrong
when inspecting test output.

The create_and_determine() helper function for ImageFormat tests have to be
removed since ASSERT_TRUE() can not be used in a function that returns a
value.