Use Result for command_run()
Rename CommandResult struct to CommandOutput. Result<CommandResult> is not
nice.

ChildProcess helper class is removed. Not really needed to make sure child
is reaped. It was used to store exit status though since a bool was
returned from wait() to indicate success or failure. Both success and exit
code can be returned with a Result<int> so "main" reason for ChildProcess
to exist is gone.