Have API for reading PLY files/data be functions returning Result
PLYReader class will be removed completely eventually.

There is a significant slowdown when loading Lucy if PLYReader is put in
an anonymous namespace. Not sure why. GCC shenanigans. Tried to profile and
investigate generated assembly but nothing really stands out in all the
noise. Add a TODO for investigating this more later (as mentioned above,
PLYreader will eventually be removed, will optimize BinaryReader some more
etc. so do not want to care about this too much at the moment) and
workaround it by naming the namespace. Seems to give more or less same
performance as before. Recently optimized PLYReader (6.9s->2.3s when
loading scene with Stanford Lucy and do not want to regress). A bit
weird... but it is how GCC 9.3.0 behaves on my machine. Not sure how
Clang behaves since current version available in Arch (9.0.1) is not
capable of compiling Rayni due to use of some C++20 features that are
hopefully implemented in Clang 10 that will be available soon.