Remove TextReader::EOFException
By checking buffer_position_ before using it (through at_eof()).

Also removes relying on buffer_ pointing to a '\0' terminated string to
function correctly at eof. This was not the case when reading a file that
is just mmap:ed (it was OK when std::getline() was used to read into a
std::string on a per line basis, not after MemoryMappedFile started being
used).