Remove use of std::string_view::substr() when splitting strings
It throws if position is out of range. Position passed in this case can
never be out of range though, unless find() is not implemented correctly.
Use std::string_view() constructor instead that takes pointer and length.
A bit ugly but there is no other way to prevent code that can throw
exception being generated, if std::string_view is to be used.

Noticed when looking at generated assembly for optimizing Wavefront .obj
reading. This change only results in ~10ms speedup when reading Stanford
Buddah, Dragon and Bunny (<1%), but is nevertheless worth it, I think
(unnecessary exception code is... unnecessary...).

Rename position to pos to make code fit on single line and rework it to use
less ternary conditional expressions. Was less clear without this change
when substr() was removed... perhaps...