Add functions to get next floating point value of real_t
Add global static assert in math.h that IEEE 754 floats are used since
real_next_*() depends on it.

Could use std::memcpy() instead of unions when converting from float to
integer. Using unions in this way is not standard compliant and
std::memcpy() optimizes to the same thing (register copy). But, on the
other hand, do not know of any compiler that does not do what is wanted
here...

Will be used in following commit to add function for generating ray that
avoids self intersection with object at origin due to rounding error.