Put L1 cache line size constant in config.h
Have hardcoded values in a couple of places, will change to this global
value instead.

Perhaps not necessary to read value from sysfs. More or less 64 everywhere
that matters (most of today's x86 CPUs). Could just hard code it to 64,
but tried sysfs cat and it worked, so just leave it.

Also not interested in detecting this run-time since want to use it with
alignas(). Having it be compile time does not matter for Rayni use case
(compiled with other CPU specific flags etc.).