Fix Clang spuriously warning when comparison operator is rewritten to <=>
This started happening in 9, I think, but I just ignored it since I
thought it would be fixed. It is still an issue with Clang 11 though.
Add pragmas to disable locally for now, may have to remove warning
completely for Clang if it starts occurring in more places before it is
fixed.