Change placement of opening braces for if/while etc
Just stylistic. Have not preferred style in Rayni for a very long time
(and not before Rayni either, something that was tried out at start of
project) but refrained from changing it thinking it would be too
invasive. But changed and turned out to not be that bad, so just do it.