Remove Variant::Exception
Failing to e.g. convert a Variant should not be something that
terminates the application => exceptions should not be used.

Reson for it being added was convenience (less to write in e.g.
constructors that take const Variant &), but exception use has
spread unnecessarily. Will disable them completely in following
commit, in part because of they are used "wrong" but also due to
performance.