Remove use of std::system_error exception in src/lib/system
And make it so that MainLoop is exited with an error exit code if an
error occurs. Log Error with log_error() as well.

Not sure if the GSource changes in the GTK application are correct. Think
prepare() callback has to be implemented now as well in case MainLoop()
fails. Documentation is vague though, have to check Glib source, so punt
for now.