Use string_printf() when formatting duration strings
Not sure why, but for some reason , is used instead of . for decimal
separator (in my session) when using snprintf but not streams in GTK/Qt
applications. Not when running tests though. Why?!?!? GTK/Qt seems to
do something that affects locale for snprintf() but not streams. *Sigh*
Just hack it away. Always want a '.'.